Refactor commands into helper utils files
