+++ cypress/integration/grid/keyUserJourneys.spec.js
@@ -3,11 +3,11 @@ import 'cypress-file-upload';
 
 import { getDomain, setCookie } from '../../utils/networking';
 import { checkVars } from '../../utils/vars';
-import {
-  deleteImages,
-  getImageHash,
-  getImageURL,
-} from '../../utils/grid/image';
+import { deleteImages, getImageHash, getImageURL } from '../../utils/grid/api';
+import * as uploads from '../../utils/grid/upload';
+import * as crops from '../../utils/grid/crop';
+import * as image from '../../utils/grid/image';
+import * as collections from '../../utils/grid/collections';
 const config = require('../../../env.json');
 
 // ID of `cypress/fixtures/GridmonTestImage.png`
@@ -90,38 +90,10 @@ describe('Grid Key User Journeys', function () {
         'Usage rights before rights are added'
       ).to.equal('{}');
 
-      // Set rights as screengrab
-      cy.get('ui-upload-jobs [data-cy=edit-rights-button]')
-        .click({ force: true })
-        .get('ui-upload-jobs [data-cy=it-rights-select]')
-        .select('screengrab')
-        .get('ui-upload-jobs [data-cy=it-edit-usage-input]')
-        .type(Date.now().toString())
-        .get('ui-upload-jobs [data-cy=save-usage-rights]')
-        .click()
-        .should('not.exist');
-
-      // Add label
-      cy.get('ui-upload-jobs [data-cy=it-add-label-button]')
-        .click()
-        .get('ui-upload-jobs [data-cy=label-input]')
-        .type('integration-test-label')
-        .get('ui-upload-jobs [data-cy=save-new-label-button]')
-        .click()
-        .should('not.exist');
-
-      // Add credit
-      cy.get('ui-upload-jobs [data-cy=image-metadata-credit]')
-        .clear()
-        .type('Editorial Tools Integration Tests');
-
-      // Add image to collection
-      cy.get('ui-upload-jobs [data-cy=add-image-to-collection-button]').click();
-      cy.get('.collection-overlay__collections')
-        .find('[data-cy="Cypress Integration Testing-collection"]')
-        .contains('Cypress Integration Testing')
-        .click({ force: true })
-        .should('not.exist');
+      uploads.setRights('screengrab', date);
+      uploads.addLabel('integration test label');
+      uploads.addCredit('Editorial Tools Integration Tests');
+      uploads.addImageToCollection('Cypress Integration Testing');
 
       cy.get(`ui-upload-jobs [href="/images/${dragImageID}"] img`).click();
       cy.url()
@@ -141,18 +113,13 @@ describe('Grid Key User Journeys', function () {
     // Wait for cropper image to exist before continuing
     cy.get('.cropper-face').should('exist');
 
-    // Select freeform crop
-    cy.get('[data-cy=crop-options]').contains('freeform').click();
-    // Edit width
-    cy.get('[data-cy=crop-width-input]').clear().type(crop.width);
-    // Edit height
-    cy.get('[data-cy=crop-height-input]').clear().type(crop.height);
-    // Edit x coordinate
-    cy.get('[data-cy=crop-x-value-input]').clear().type(crop.xValue);
-    // Edit y coordinate
-    cy.get('[data-cy=crop-y-value-input]').clear().type(crop.yValue);
-
-    cy.get('.button').click().wait(waits.createCrop);
+    crops.createCrop(
+      crop.width,
+      crop.height,
+      crop.xValue,
+      crop.yValue,
+      waits.createCrop
+    );
 
     cy.url({ timeout: 5000 }).should(
       'equal',
@@ -168,84 +135,28 @@ describe('Grid Key User Journeys', function () {
       ).to.be.greaterThan(0);
     });
 
-    // Delete all crops
-    cy.get('[data-cy=delete-all-crops-button]')
-      .click()
-      .click()
-      .get('[data-cy=delete-all-crops-button]')
-      .should('not.exist')
-      .then(async () => {
-        const cropsAfterDelete = (await axios.get(cropsUrl)).data.data;
+    crops.deleteAllCrops();
 
-        expect(
-          cropsAfterDelete.length,
-          'Number of crops after delete ALL'
-        ).to.equal(0);
-      });
+    cy.then(async () => {
+      const cropsAfterDelete = (await axios.get(cropsUrl)).data.data;
+
+      expect(
+        cropsAfterDelete.length,
+        'Number of crops after delete ALL'
+      ).to.equal(0);
+    });
   });
 
   it('User can edit the image rights, description, byline, credit, copyright, label', () => {
     cy.visit(getImageURL()).wait('@getImage');
 
-    //  Edit the rights
-    cy.get('[data-cy=it-edit-usage-rights-button]').click({ force: true });
-    cy.get('[data-cy=it-rights-select]').select('screengrab');
-    cy.get('[data-cy=it-edit-usage-input]').type(date);
-    cy.get('.ure__bar > .button-save')
-      .click({ timeout: 5000 }) // Why do we need to wait?
-      .should('not.exist');
-
-    // Edit the description
-    cy.get('[data-cy=it-edit-description-button]').click({ force: true });
-    cy.get('[data-cy=metadata-description] .editable-has-buttons')
-      .clear()
-      .type(date);
-    cy.get('[data-cy=metadata-description] .editable-buttons > .button-save')
-      .click()
-      .should('not.exist');
-
-    // Edit the byline
-    cy.get('[data-cy=it-edit-byline-button]').click({ force: true });
-    cy.get('[data-cy=metadata-byline] .editable-has-buttons')
-      .clear()
-      .type(date);
-    cy.get('[data-cy=metadata-byline] .editable-buttons > .button-save')
-      .click()
-      .should('not.exist');
-
-    // Edit the credit
-    cy.get('[data-cy=it-edit-credit-button]').click({ force: true });
-    cy.get('[data-cy=metadata-credit] .editable-has-buttons')
-      .clear()
-      .type(date);
-    cy.get('[data-cy=metadata-credit] .editable-buttons > .button-save')
-      .click()
-      .should('not.exist');
-
-    // Edit the copyright
-    cy.get('[data-cy=it-edit-copyright-button]').click({ force: true });
-    cy.get('[data-cy=metadata-copyright] .editable-has-buttons')
-      .clear()
-      .type(date);
-    cy.get('[data-cy=metadata-copyright] .editable-buttons > .button-save')
-      .click()
-      .should('not.exist');
-
-    // Add label
-    cy.get('[data-cy=it-add-label-button]').click();
-    cy.get('.text-input').clear().type('someLabelHere');
-    cy.get('.gr-add-label__form__buttons__button-save')
-      .click()
-      .should('not.exist');
-    cy.get('.labeller')
-      .contains('someLabelHere', { timeout: 5000 })
-      .should('exist');
-    cy.get('.labeller')
-      .contains('someLabelHere')
-      .parent()
-      .find('[data-cy=it-remove-label-button]')
-      .click()
-      .should('not.exist');
+    image.editRights('screengrab', date);
+    image.editDescription(date);
+    image.editByline(date);
+    image.editCredit(date);
+    image.editCopyright(date);
+    image.addLabel(date);
+    image.removeLabel(date);
   });
 
   it('User can create a child collection and delete it', () => {
@@ -257,55 +168,15 @@ describe('Grid Key User Journeys', function () {
     // Click on collections panel
     cy.get('[data-cy=show-collections-panel]').should('exist').click();
 
-    // Click on edit collections button
-    cy.get('[data-cy=edit-collections-button]').click();
-
-    // Click on button create new child
-    cy.get(`[data-cy="${collectionName}-collection"]`)
-      .find('[data-cy=create-new-folder-button]')
-      .click({ force: true });
-
-    // Type name of new child
-    cy.get(`[data-cy="${collectionName}-collection"]`)
-      .parent()
-      .find('[data-cy=collection-child-input]')
-      .type(childName);
-
-    // Save child
-    cy.get(`[data-cy="${collectionName}-collection"]`)
-      .parent()
-      .find('[data-cy=save-child-button]')
-      .click({ force: true })
-      .should('not.exist');
-
-    // Stop editing collections
-    cy.get('[data-cy=edit-collections-button]').click();
-
-    // Go to new collection
-    cy.get(`[data-cy="${collectionName}-collection"] button`).click({
-      force: true,
-    });
-    cy.get(`[data-cy="${collectionName}-collection"]`)
-      .parent()
-      .contains(childName)
-      .click({ force: true });
+    collections.createChild(collectionName, childName);
+    collections.goToChild(collectionName, childName);
 
     cy.get('.search-query').should(
       'contain',
       `${collectionName.toLowerCase()}/${childName}`
     );
 
-    // Click on edit collections button
-    cy.get('[data-cy=edit-collections-button]').click();
-
-    // Delete child collection
-    cy.get(`[data-cy="${collectionName}-collection"]`)
-      .parent()
-      .contains(childName)
-      .parent()
-      .contains('delete')
-      .click({ force: true })
-      .click({ force: true });
+    collections.deleteChild(collectionName, childName);
 
     // Assert collection does not exist
     cy.get(`[data-cy="${collectionName}-collection"]`)
