+++ cypress/utils/grid/collections.js
@@ -1,0 +1,52 @@
+module.exports = {
+  createChild(collection, name) {
+    // Click on edit collections button
+    cy.get('[data-cy=edit-collections-button]').click();
+
+    // Click on button and create new child
+    cy.get(`[data-cy="${collection}-collection"]`)
+      .find('[data-cy=create-new-folder-button]')
+      .click({ force: true });
+
+    // Type name of new child
+    cy.get(`[data-cy="${collection}-collection"]`)
+      .parent()
+      .find('[data-cy=collection-child-input]')
+      .type(name);
+
+    // Save child
+    cy.get(`[data-cy="${collection}-collection"]`)
+      .parent()
+      .find('[data-cy=save-child-button]')
+      .click({ force: true })
+      .should('not.exist');
+
+    // Stop editing collections
+    cy.get('[data-cy=edit-collections-button]').click();
+  },
+
+  goToChild(collection, child) {
+    // Go to new collection
+    cy.get(`[data-cy="${collection}-collection"] button`).click({
+      force: true,
+    });
+    cy.get(`[data-cy="${collection}-collection"]`)
+      .parent()
+      .contains(child)
+      .click({ force: true });
+  },
+
+  deleteChild(collection, child) {
+    // Click on edit collections button
+    cy.get('[data-cy=edit-collections-button]').click();
+
+    // Delete child collection
+    cy.get(`[data-cy="${collection}-collection"]`)
+      .parent()
+      .contains(child)
+      .parent()
+      .contains('delete')
+      .click({ force: true })
+      .click({ force: true }); // confirm delete crop
+  },
+};
