+++ cypress/utils/grid/crop.js
@@ -1,0 +1,25 @@
+module.exports = {
+  createCrop(width, height, xValue, yValue, wait) {
+    // Select freeform crop
+    cy.get('[data-cy=crop-options]').contains('freeform').click();
+    // Edit width
+    cy.get('[data-cy=crop-width-input]').clear().type(width);
+    // Edit height
+    cy.get('[data-cy=crop-height-input]').clear().type(height);
+    // Edit x coordinate
+    cy.get('[data-cy=crop-x-value-input]').clear().type(xValue);
+    // Edit y coordinate
+    cy.get('[data-cy=crop-y-value-input]').clear().type(yValue);
+
+    cy.get('.button').click().wait(wait);
+  },
+
+  deleteAllCrops() {
+    // Delete all crops
+    cy.get('[data-cy=delete-all-crops-button]')
+      .click()
+      .click()
+      .get('[data-cy=delete-all-crops-button]')
+      .should('not.exist');
+  },
+};
