+++ cypress/utils/grid/image.js
@@ -1,56 +1,68 @@
-import axios from 'axios';
-import { getDomain, setCookie } from '../networking';
-// hash of the image in assets/GridmonTestImage.png
-export const imageHash = 'fe052e21c4bc4d76a2c841d97c5b2281cccd19bd';
+module.exports = {
+  editRights(rightsType, usage) {
+    cy.get('[data-cy=it-edit-usage-rights-button]').click({ force: true });
+    cy.get('[data-cy=it-rights-select]').select(rightsType);
+    cy.get('[data-cy=it-edit-usage-input]').type(usage);
+    cy.get('.ure__bar > .button-save')
+      .click({ timeout: 5000 }) // Why do we need to wait?
+      .should('not.exist');
+  },
 
-export function getImageHash() {
-  return imageHash;
-}
+  editDescription(description) {
+    cy.get('[data-cy=it-edit-description-button]').click({ force: true });
+    cy.get('[data-cy=metadata-description] .editable-has-buttons')
+      .clear()
+      .type(description);
+    cy.get('[data-cy=metadata-description] .editable-buttons > .button-save')
+      .click()
+      .should('not.exist');
+  },
 
-export function getImageURL() {
-  return `${getDomain()}images/${getImageHash()}`;
-}
+  editByline(byline) {
+    cy.get('[data-cy=it-edit-byline-button]').click({ force: true });
+    cy.get('[data-cy=metadata-byline] .editable-has-buttons')
+      .clear()
+      .type(byline);
+    cy.get('[data-cy=metadata-byline] .editable-buttons > .button-save')
+      .click()
+      .should('not.exist');
+  },
 
-export async function deleteImages(cy, images) {
-  setCookie(cy, false);
-  cy.then(async () => {
-    await Promise.all(
-      images.map((id) => {
-        const url = `${getDomain('api')}images/${id}`;
-        axios
-          .delete(url)
-          .catch((err) => {
-            // If it's 404, it means the image doesn't exist (so it can't be deleted)
-            if (err.response && err.response.status === 404) {
-              cy.log(`${id} doesn't exist in Grid`);
-              return;
-            }
-            cy.log(`Error deleting ${id}`, err.message);
-            throw err;
-          })
-          .then((res) => expect(res.status, `Delete ${id}`).to.equal(202));
-      })
-    );
-  });
-}
+  editCredit(credit) {
+    cy.get('[data-cy=it-edit-credit-button]').click({ force: true });
+    cy.get('[data-cy=metadata-credit] .editable-has-buttons')
+      .clear()
+      .type(credit);
+    cy.get('[data-cy=metadata-credit] .editable-buttons > .button-save')
+      .click()
+      .should('not.exist');
+  },
 
-export async function uploadImage(cy, image) {
-  setCookie(cy);
-  const url = `${getDomain('loader')}images`;
-  const res = await axios
-    .post(url, image, { withCredentials: true })
-    .catch(async (err) => {
-      console.error('uploadImage error', url);
-      console.error(err);
-      throw err;
-    });
-  expect(res.status, 'Upload test image').to.equal(202);
-  cy.log(`Uploaded test image was ${res.statusText}`);
-  return null;
-}
+  editCopyright(copyright) {
+    cy.get('[data-cy=it-edit-copyright-button]').click({ force: true });
+    cy.get('[data-cy=metadata-copyright] .editable-has-buttons')
+      .clear()
+      .type(copyright);
+    cy.get('[data-cy=metadata-copyright] .editable-buttons > .button-save')
+      .click()
+      .should('not.exist');
+  },
 
-export function readAndUploadImage(cy) {
-  cy.task('readFileMaybe', 'assets/GridmonTestImage.png').then(
-    async (contents) => await uploadImage(cy, Buffer.from(contents.data))
-  );
-}
+  addLabel(name) {
+    cy.get('[data-cy=it-add-label-button]').click();
+    cy.get('.text-input').clear().type(name);
+    cy.get('.gr-add-label__form__buttons__button-save')
+      .click()
+      .should('not.exist');
+    cy.get('.labeller').contains(name, { timeout: 5000 }).should('exist');
+  },
+
+  removeLabel(date) {
+    cy.get('.labeller')
+      .contains(date)
+      .parent()
+      .find('[data-cy=it-remove-label-button]')
+      .click()
+      .should('not.exist');
+  },
+};
